UPSTREAM: Makefile.inc: Export CCACHE_EXTRAFILES and add __BUILD_DIR__ definition
This patch adds the CCACHE_EXTRAFILES variable to the list of exported
environment variables, which can be useful as a target-specific variable
to make ccache aware of extra dependencies that it cannot figure out on
its own. It also adds a CPPFLAGS parameter to define the __BUILD_DIR__
constant for the preprocessor so that the current output build directory
can be referenced in C code if necessary.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie3bd70e0ed9fc2808ea101b6a8960618321ef4fe
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9123449734f43922fe39cdb08c3d60f02f0eb3ed
Original-Change-Id: I4fdd08842972cfed8ef5e5a61ebf859c0571bcfb
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/26451
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1068845
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>