UPSTREAM: mb/google/kahlee: Init APU_BIOS_FLASH_WP_L GPIO to reset stage
GPIO APU_BIOS_FLASH_WP_L is first read in ROM stage to determine the
state of the BIOS FLASH Write Protect signal at boot.
The result of this read accumulated in the vboot state that's passed on
to the upper layers of the stack.

Therefore this GPIO must be configured as a "reset stage" GPIO, not
a "RAM" stage GPIO.

BUG=b:79866233
TEST=firmware_WriteProtect

Change-Id: I8bf38d07c0b566b70785110cc7707d7eb745c5e1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c3e7416b1bae8dfa6f3b43031de78a0823deee4f
Original-Change-Id: I1d96ab4bbfeaf9db9f74cf0c58cbab2104079bf7
Original-Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/26498
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1073280
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Trybot-Ready: Martin Roth <martinroth@chromium.org>