UPSTREAM: i210: Add additional PCI-ID to the i210 driver
When the i210 MACPHY is operated in the SERDES Backplane mode (which
depends on the programmed firmware image), its PCI-ID will be 0x1537.
This does however not change the programming interface for the MAC
address.
Therefore add this new PCI-ID to the driver so that the MAC address can
be programmed in this operation mode as well.

BUG=none
BRANCH=none
TEST=none

Change-Id: I940f2812d8955b4003cd76ec947f2673d872490b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7731932937b2f0f4f968a6c8c3b2ebb4429de034
Original-Change-Id: I608535202c49e40690381c2b2ab26322d62cfb37
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/26683
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Reviewed-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1084688
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>