UPSTREAM: mb/google/octopus: Fix invalid IOSSTATE settings for 1.8V pins
When the normal termination is None, the standby termination is
none also as per Doc# 572688. So when termination is only needed in
standby, use the IOSSTATE setting that drives low/high via the
Tx mode instead.

Also disabled Speaker in Standby state to save power.

BUG=b:79874891, b:79982669
BRANCH=None
TEST=Compiled and flashed image on Bip. Checked that suspend_resume cycles
pass. Checked that bluetooth is functional on resume. On Yorp, checked
that speaker is functional after a suspend/resume cycle.

Change-Id: Ie8de7b9046007c16a8e32d70ddf3e12f828f4df9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 280bc303460ea400d2dbb2590a4e68390814c1ee
Original-Change-Id: I6a3852548f944176a80feb32e9885b03b8af25db
Original-Signed-off-by: Shamile Khan <shamile.khan@intel.com>
Original-Reviewed-on: https://review.coreboot.org/26762
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1088685
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>