UPSTREAM: google/fizz: fix LAN driver chip_info attachment
As a result of commit:
[711fb81] soc/intel/skylake: Swap PCI devfn resides in same PCI device

fizz's chip_info for the LAN driver is being overwritten/nulled, as the
LAN device is on function 2 (PCIe port 3), but the driver info was set
for the post-swapped PCIe port (1).
Move the driver chip_info to function 2/port 3, so that it follows the
PCI device function when swapped after FSP-s, and is correctly passed
to the LAN driver.

Test: boot google/fizz (teemo variant), check cbmem console and
verify ethernet MAC address and LED config correctly set.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic5ef1c9e4b17718b4026d1bf668922ac81337587
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f061a73ab9654ccae73b2fcbd58d8f08017bae59
Original-Change-Id: I08810c0c89d99af5799f42c7c4e51814f09aafec
Original-Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/27012
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1095092
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>