UPSTREAM: mb/*/*: Harmonise FD and devicetree on boards featuring ICH7
On some boards the devicetree and Function Disable register did not
match. In this case the FD values are put in the devicetree as these
were the values that were actually used in practice.

A complete devicetree will make it easier to automatically disable
devices in ramstage.

BUG=none
BRANCH=none
TEST=none

Change-Id: I34d77c904a6ef990599a967ca98111a6cd317b9b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b9d2589ca40026b543ecb5b008ce0d1bc346bf53
Original-Change-Id: I1692ca5f490ea84e2fc520d3f66044ad7514f76e
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/27122
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1341314
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>