UPSTREAM: Kconfig: Add config to insert ucode address in second FIT
This config is used to provide the name of a region where a microcode
is located. The address of this will be added as the first entry in
the FIT of the topswap bootblock.

This adds a capability to associate two microcodes for each
of the two bootblocks, this allows for the CPU to boot with different
microcodes with 2 separate bootblocks.

BUG=none
BRANCH=none
TEST=none

Change-Id: I85d05eb90d857028700a4b0d7921190d4a2392db
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 53485124513a76ebe85a80f59e8dff9915f518bd
Original-Change-Id: I4ee41d90bae34862aa68c9b8bd69288de1335585
Original-Signed-off-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Original-Reviewed-on: https://review.coreboot.org/27151
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1143259
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>