UPSTREAM: fsp_broadwell_de: Add uncore ASL in common location
This patch adds a file for uncore ASL code that is common among
Broadwell-DE mainboards but is currently copy + pasted in each of
their dsdt.asl files.

This is only for clean-up purposes. It is unclear if the code itself is
really necessary, but until we can do further investigation and testing
it will be left in.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iecaaab0d3f9621af90612642b419751f19f7ffc8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7000f5a0c151f81b891a0c21b1ca680ff5f254af
Original-Change-Id: I188e5e46dfa7c2ed3991fb97f2c1b5e062e2212d
Original-Signed-off-by: David Hendricks <dhendricks@fb.com>
Original-Reviewed-on: https://review.coreboot.org/27155
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/1109776
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>