UPSTREAM: soc/amd/stoneyridge/southbridge.c: Fix get_index_bit limit check
Limit is the maximum number of bits to be tested, however it's being checked
against the number of bytes of uint32_t. when it should be number of bits.
Create a macro to provide the number of bits, and use it instead of sizeof.

BUG=b:75996437
TEST=Add debug messages to see code passing beyond the check, build and
boot grunt, check that it passed the limit check, remove debug code.

Change-Id: Id39cac463a1d6ade0fddb4dca910e6e613372f16
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ef73cb88dc52dbe9ea5cc23939e109e0a7967e89
Original-Change-Id: Id1dfda26d789183b346b20c37fec923d996b80db
Original-Signed-off-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-on: https://review.coreboot.org/27162
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1115230
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>