UPSTREAM: drivers/i2c: Add a driver for Semtech SX9310
This adds a new driver for the SX9310 proximity detector device.

The purpose of this is to enable the device's calibration information
to be stored in firmware, and then transferred over to the kernel
via ACPI.

This device has more than 10 individual configuration parameters,
so they would not fit in the generic driver's properties table.

BUG=none
BRANCH=none
TEST=none

Change-Id: I25813006d4389f50d583a3e880b93a2815852df3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 76a1f49f742df6d1bde38fa36f4b06a193e6e4f5
Original-Change-Id: Id8c434eec9fe2da731e142442503a12e88db2236
Original-Signed-off-by: Enrico Granata <egranata@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/27173
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1111966
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>