UPSTREAM: smm: Push a null stack base pointer
When generating a backtrace we need an indicator when we have hit the
beginning of the stack. The i386 ABI states that %ebp points to the next
stack frame. NULL can be used to indicate the end of the stack.

We could add a NULL return pointer at %ebp+4, but I decided to omit it
since a NULL stack pointer can be used as an indicator that there is no
return pointer.

BUG=b:80539294
TEST=built and tested on grunt

Change-Id: I5834c664736aa30830cef0732d8cae9e150dd7a3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 871156898c874c5247d4024309af348f8bc1e518
Original-Change-Id: I8a48114d31a5c716335d264fa4fe4da41dc5bf11
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/27226
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1118460
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>