UPSTREAM: nb/intel/gm45: Don't use PCI operations on the pci_domain device
pci ops happen to work on this struct device since the device_path is an union.

This patch still keeps adding the fixed resources in the pci_domain
ops since moving it to the PCI ops which could properly use the
function argument for PCI operations would require all PCI IDs to be
added or else breakages are to be expected.

BUG=none
BRANCH=none
TEST=none

Change-Id: I440bcf19f30e0cff141c7b0cc51de860c6ca4b3b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8908931f1ea458f0482265583746788589506380
Original-Change-Id: I598b056d5fb1ce23b390b2f0ab4e9fb242d3685a
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/27242
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1159009
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>