UPSTREAM: nb/intel/pineview: Don't use PCI operations on the pci_domain device
pci ops happen to work on this struct device since the device_path is an union.

This patch still keeps adding the fixed resources in the pci_domain
ops since moving it to the PCI ops which could properly use the
function argument for PCI operations would require all PCI IDs to be
added or else breakages are to be expected.

BUG=none
BRANCH=none
TEST=none

Change-Id: I37d52906624aed6e83c2731704f4ac2f21eb7cdd
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 15e1b39e6e4c7d7d7b0bc2bc7347a1f037ec2692
Original-Change-Id: Iea5a09c62cca102b2c211e9256295c24cf3e9fa0
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/27243
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1159008
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>