UPSTREAM: nb/intel/x4x: Don't use PCI operations on the pci_domain device
pci ops happen to work on this struct device since the device_path is an union.

This patch still keeps adding the fixed resources in the pci_domain
ops since moving it to the PCI ops which could properly use the
function argument for PCI operations would require all PCI IDs to be
added or else breakages are to be expected.

BUG=none
BRANCH=none
TEST=none

Change-Id: If03364b0f97308cfc1d352def2a86c7a7cdaddb2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c6e13b6690e15fe32b7c7c135bea5e8a7e6d1753
Original-Change-Id: Iabfd15884ec8feb846d01b6af3c4afe5c1494feb
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/27245
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1159007
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>