UPSTREAM: riscv: add trampoline in MBR block to support boot mode 1
Add "j pc + 0x0800" at the beginning of the MBR to jump to bootblock.

Tested on hardware:
boot mode 15: works as before
boot mode 1: jump to bootblock works, but bootblock needs to be modified to
move the stack to L2LIM. This will be in a separate commit.

Further changes are needed in the bootblock

BUG=none
BRANCH=none
TEST=none

Change-Id: I5f7109b3d5bb06d9240d1d7884e2314087a26aad
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2326a284ac6a6646a918331425952ece2da723c1
Original-Change-Id: I16e762d9f027346b124412f1f7ee6ff37f431d86
Original-Signed-off-by: Philipp Hug <philipp@hug.cx>
Original-Reviewed-on: https://review.coreboot.org/27397
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jonathan Neuschfer <j.neuschaefer@gmx.net>
Reviewed-on: https://chromium-review.googlesource.com/1227769
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>