UPSTREAM: what-jenkins-does: Pass V=1 through to abuild
Even though we were setting V=1 in the build, this wasn't getting passed
to abuild, so the builds there didn't have additional debug information.
That made it difficult to debug issues on the builder. This sets the
verbose flag for abuild if V=1 is set.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0fd8874f740f8b1cfdc96c7ef1c1a417fa5c98f7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6d4a060bba907ca29b2c69fda9a4ae3e3c73fc44
Original-Change-Id: Id9ec50add9693a6c36ffdb5c78d148d0fc012549
Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/27492
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1143245
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>